% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_sampler.R
\name{runTwoStateGibbs}
\alias{runTwoStateGibbs}
\title{Gibbs sampler for a phylogenetic model with a binary trait.}
\usage{
runTwoStateGibbs(inputTrees, rootDist, traitData, initLambda01, initLambda10,
  priorAlpha01, priorBeta01, priorAlpha10, priorBeta10, mcmcSize, mcmcBurnin,
  mcmcSubsample)
}
\arguments{
\item{inputTrees}{\code{multiPhylo} object containing a list of trees.}

\item{rootDist}{Numeric vector of size two with probabilities of the 
two states at the root of the tree.}

\item{traitData}{Vector of trait values, coded as 0 and 1.}

\item{initLambda01}{Initial value of the 0 -> 1 rate.}

\item{initLambda10}{Initial value of the 1 -> 0 rate.}

\item{priorAlpha01}{Shape parameter of the Gamma prior on the 0 -> 1 rate.}

\item{priorBeta01}{Rate parameter of the Gamma prior on the 0 -> 1 rate.}

\item{priorAlpha10}{Shape parameter of the Gamma prior on the 1 -> 0 rate.}

\item{priorBeta10}{Rate parameter of the Gamma prior on the 0 -> 1 rate.}

\item{mcmcSize}{Total number of MCMC iterations.}

\item{mcmcBurnin}{Number of initial iterations to ignore.}

\item{mcmcSubsample}{Integer controlling how many MCMC iterations to save. 
For example, \code{mcmcSubsample = 10} saves every tenth iteration after ignoreing the first \code{mcmcBurnin} iterations.}
}
\value{
List with two elements: the input trees and a matrix of MCMC output.
The columns of MCMC output matrix are Iteration number, Index of a tree 
that was sampled at this iteration, Log posterior, 
0 -> 1 rate, 1 -> 0 rate, Number of 0 -> 1 jumps, Number of 1 -> 0 jumps,
Time spent in state 0, Time spent in state 1
}
\description{
Runs a Gibbs sampler for a two-state CMTC 
on a set of phylogenetic trees, with data 
at the tips of the trees.
}
